Add "Folder" table and update schema relationships
This migration introduces a new "Folder" table and modifies existing database relationships. The "File" table is updated with "createdAt" and "folderId" columns, and its "extension" column is made optional. Additionally, the "member" column is removed from the "User" table, and necessary foreign key constraints are added to ensure relational integrity.